fix cell04 ex02

+++ cell04/ex02/calculator.py
@@ -5,7 +5,7 @@ second_num = int(input("Give me the second number: "))
 print("Thank you")
 opt = ["+","-","/","*"]
 try:
-    ans = [first_num + second_num , first_num - second_num , int(first_num / second_num) , first_num * second_num]
+    ans = [first_num + second_num , first_num - second_num , first_num / second_num , first_num * second_num]
     for i in range(0,4):
         print(f"{first_num} {opt[i]} {second_num} = {ans[i]}")
 except:
